requirements/requirements.txt
langchain==0.1.0
openai==1.3.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
streamlit==1.28.0
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
transformers==4.35.0
torch==2.1.0 